- Added experimental support for extracting YouTube comments with `yt-dlp`. (disabled by default)
Added via the config-options:

	    ytdlp_comments
	    ytdlp_max_comments
	    ytdlp_max_comment_depth

ytdlp_comments: when set to `1`, use `yt-dlp` for extracting YouTube comments.
ytdlp_max_comments: maximum number of comments to extract with `yt-dlp` (replies count as comments).
ytdlp_max_comment_depth: depth for YouTube comments: 1 (without replies) or 2 (with replies).

Loading more comments on-demand, is currently not supported.

On the CLI version, the following command-line options are also available:

	--ytdlp-comments!
	--max-comments=i
	--comments-depth=i

Thanks to @pirate486743186 for info about the usage of `yt-dlp` for extracting YouTube comments!