- Display likes and dislikes in video info.
In the CLI verison, the likes and dislikes are displayed in `:i=n` or when selecting a video for playing.
In the GTK version, the likes and dislikes are displayed in the "Extra details" window.

This is done by fetching the `/watch?v=ID` page and extracting the number of likes. Then, from `youtubei/player` we get the average rating of the video, from which we compute the approximate number of dislikes.