gui: simplify how a number of child windows are handled
Use a dialog, instead of another top-level window, so "escape to close"
is automatically supported, and the close button can be simplified to
use the corresponding response.

Use a generic hide/show callback to avoid the need for a bunch of
specific ones.

If there are corresponding menu items, just set our shortcuts for
showing each dialog there.

Incidentally, this fix a Gtk warning when closing the errors window:

```
FATAL: could not convert value 0 to enum type GtkResponseType at […]/Gtk3.pm line 285.
```